% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso_py.R
\name{lasso}
\alias{lasso}
\title{lasso() function}
\usage{
lasso(form, df, contrasts = NULL, lambda)
}
\arguments{
\item{form}{a formula;}

\item{df}{a data frame used for the function;}

\item{contrasts}{a list of contrasts for factor variables}

\item{lambda}{a constant for penalty}
}
\description{
to build a lasso regression model using python code
}
\examples{
data(iris)
library(reticulate)
library(casl)
fit_model <- lasso(Sepal.Length ~ ., iris, contrasts = list(Species = "contr.sum"), lambda = 0.1)
}
