% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_ofc.R
\name{lm_ofc}
\alias{lm_ofc}
\title{lm_ofc() function}
\usage{
lm_ofc(form, df, contrasts = NULL)
}
\arguments{
\item{form}{a formula;}

\item{df}{a data frame used for the function;}

\item{contrasts}{a list of contrasts for factor variables}
}
\description{
to build a linear regression model that reads in contiguous rows from a data frame
}
\examples{
data(iris)
library(reticulate)
fit_linear_model <- lm_ofc(Sepal.Length ~ ., iris, contrasts = list(Species = "contr.sum"))
}
