% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridge_py.R
\name{ridge_py}
\alias{ridge_py}
\title{ridge() function}
\usage{
ridge_py(form, df, contrasts = NULL, lambda = 0)
}
\arguments{
\item{form}{a formula;}

\item{df}{a data frame used for the function;}

\item{contrasts}{a list of contrasts for factor variables}

\item{lambda}{a constant for penalty}
}
\description{
to build a ridge regression model using python code
}
\examples{
data(iris)
library(reticulate)
fit_linear_model <- ridge_py(Sepal.Length ~ .- Species, iris, contrasts = list(Species = "contr.sum"), lambda = 0.1)
}
